pyproject.toml
[project]
name = "mlops-demo-iris"
version = "0.1.0"
description = "Minimal Iris MLOps demo with FastAPI, scikit-learn, GCS, and Cloud Run."
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "scikit-learn==1.5.2",
    "joblib==1.4.2",
    "google-cloud-storage==2.18.2",
    "python-dotenv==1.0.1",
    "numpy==1.26.4",
]

[tool.uv]
index-url = "https://pypi.org/simple"

